PYIC-4056: Move CriStoringService to a lib
The logic in the CriStoringService can be re-used for the ticf CRI
lambda, so this moves it out of the process-cri-callback lambda and to a
shared lib.
