+++ lambdas/process-cri-callback/build.gradle
@@ -15,15 +15,17 @@ dependencies {
 			"com.amazonaws:aws-lambda-java-core:$rootProject.ext.dependencyVersions.awsLambdaJavaCore",
 			"com.amazonaws:aws-lambda-java-events:$rootProject.ext.dependencyVersions.awsLambdaJavaEvents",
 			"com.nimbusds:oauth2-oidc-sdk:$rootProject.ext.dependencyVersions.nimbusdsOauth2OidcSdk",
+			project(":libs:audit-service"),
+			project(":libs:cimit-service"),
 			project(":libs:common-services"),
+			project(":libs:cri-response-service"),
+			project(":libs:cri-storing-service"),
+			project(":libs:gpg45-evaluator"),
 			project(":libs:journey-uris"),
-			project(":libs:audit-service")
-	implementation project(path: ':libs:kms-es256-signer')
-	implementation project(path: ':libs:cimit-service')
-	implementation project(path: ':libs:cri-response-service')
-	implementation project(path: ':libs:verifiable-credentials')
-	implementation project(path: ':libs:gpg45-evaluator')
-	implementation project(path: ':libs:user-identity-service')
+			project(":libs:kms-es256-signer"),
+			project(":libs:user-identity-service"),
+			project(":libs:verifiable-credentials")
+
 	testImplementation 'junit:junit:4.13.1'
 
 	aspect "software.amazon.lambda:powertools-logging:$rootProject.ext.dependencyVersions.powertoolsLogging",
