+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/dto/CriCallbackRequest.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.processcricallback.dto;
+package uk.gov.di.ipv.core.library.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
