+++ settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'di-ipv-core-back'
-include "libs:common-services",
+include "integration-test",
 		"lambdas",
 		"lambdas:build-client-oauth-response",
 		"lambdas:build-cri-oauth-request",
@@ -15,13 +15,14 @@ include "libs:common-services",
 		"lambdas:process-cri-callback",
 		"lambdas:process-journey-event",
 		"lambdas:reset-identity",
-		"integration-test",
-		"libs:kms-es256-signer",
-		"libs:verifiable-credentials",
-		"libs:journey-uris",
-		"libs:cimit-service",
-		"libs:cri-response-service",
 		"libs:audit-service",
+		"libs:cimit-service",
+		"libs:common-services",
+		"libs:cri-response-service",
+		"libs:cri-storing-service",
 		"libs:gpg45-evaluator",
+		"libs:journey-uris",
+		"libs:kms-es256-signer",
 		"libs:user-identity-service",
+		"libs:verifiable-credentials",
 		"local-running"
